qmlls: move Location computation logic to utils
It turns out that QQmlJSUtils::Location is missing an end location
which is actually needed by the LSP.

The qmlls modules worked around that by manually creating the end
location from the source code, which is hard to test, mainly untested
and buggy. In this case, qqmllsutils would return a correct Location,
and the qmlls modules would fail on loading the qmlfile to get the
missing position information while trying to convey the Location to
the LSP layer.

Therefore, extend Location to actually contain the end position and move
the end position computations to qqmllsutils which has much better test
coverage because it is way easier to test.

Extend the Location computation logic by adding a TextPosition field
in QQmlJSUtils::Location and adding adequate static from() and tryFrom()
construction methods in Location. Make the fields private and add some
getters to avoid the newly added field from being forgotten.

This is needed for QTBUG-127661 because you can't retrieve the qml file
code from the DomItem if the filepath points to the source folder
instead of the build folder, except if you happened to open the build
folder file in your editor.

Pick-to: 6.7
Task-number: QTBUG-127661
Change-Id: Id2ff400feeebeb90da6cfeadbd81260ac99ce46a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 8d874a1b3ca3f5f6db48a526ca62215464f97ef5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>