QQmlAdapterModel: guard against item deletion during notification
Complements 75ba1ce9114e320cccfbc0c14dd32675ce2e598e.
This time we also handle the case where the item was destroyed
while its own signal was being activated. So, now it won't crash
on the next signal invocation or on the item->modelDataChanged()
call.

This potentially should fix some of the hard-to-reproduce crashes
we see in users' reports. At least this definitely fixes the case
given in QTBUG-110451.

Fixes: QTBUG-110451
Pick-to: 6.8 6.7
Change-Id: I53bdb206cec8523769632b23f1546ad182f631dd
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>