Explicitly set the popupType: Popup.Item in empty model ComboBox test
ComboBox::test_emptyPopupAfterModelCleared currently tests that the
popup of a ComboBox is empty by checking the height is zero. This does
not apply to popups with popupType set to Popup.Window as they can not
have zero dimensions. Note, in the future the plan is to set a minimum
height for the ComboBox popup and so this test will become obselete.

Task-number: QTBUG-126553
Change-Id: I34e256ec86de9c722894e21d289eddbb2bab3b1a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 29faf67f7483761727f652c64374586208c25f25)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>