Baseline test: hide the cursor by using an non-visual delegate
Setting cursorVisible to false is flaky as the property is controlled by
focus handling, so we frequently get mismatches if focus is transferred
to a text input element after the cursorVisible property was set to
false.

Instead, stabilize the test by using a plain (and invisible) Item as the
cursorDelegate.

Change-Id: I5271840df106383d28f46c157610e8f3b1cab230
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 380533a494335d72224c04d2ddfa67e81ef7b32c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>