Compiler: Track all_aotstats dependencies with targets insteads of files
When invoking the all_aotstats target, all the aotstats are aggregated
or generated recursively for each module into one global set of
aotstats. The global target needs to depend on the modules so that their
stats can be aggregated.

Previously, the dependencies were specified as being the intermediate,
module-level, .aotstats files. This caused issues on some project with
more unusual structures where invoking all_aotstats would fail due to
the intermediate files not being generated.

Therefore, specify the dependencies of the global target through custom
targets for each module's stats and make the all_aotstats depend on
them.

The actual files to pass to qmlaotstats for aggregation are now stored
as target properties of those new targets.

As a drive-by, replace the use or arg_URI with the local uri.

Fixes: QTBUG-128506
Pick-to: 6.8
Change-Id: Icf0d7cb9498f68e2fface069278c44c0705089e6
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>