Example for VectorImage
This introduces a small example of VectorImage and other ways
to display SVG content in a Qt Quick application. Its purpose is
to document the available APIs for this and explain what their
differences are.

The heart.svg is a public domain, no-attribution image which is
also used in other examples and manual tests in Qt.

Pick-to: 6.8.0 6.8
Change-Id: I457ffc01d7ea286a157a4a5bff0571098b852cef
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>