Example for VectorImage
This introduces a small example of VectorImage and other ways
to display SVG content in a Qt Quick application. Its purpose is
to document the available APIs for this and explain what their
differences are.

The heart.svg is a public domain, no-attribution image which is
also used in other examples and manual tests in Qt.

Pick-to: 6.8.0
Change-Id: I457ffc01d7ea286a157a4a5bff0571098b852cef
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 252048b9561a7b27e84bf14c84659d6fec96a58b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>