tst_qquickmenubar: Make tests pass, if we change default popupType
There were a few remaining tests, that assumed that the default
popupType would always be `Popup.Item`. This patch makes the necessary
preparations, in case we decide that `Item` shouldn't be the default
one.

Note that some tests, that should be ran with both popup types, are
still only running with `popupType: Popup.Item`. Further work is needed
in order to get them working with both popup types.

In addition, close existing popups in cleanup(), in case there are some.
This were generally the case, when executing tests on Windows.

Task-number: QTBUG-126553
Pick-to: 6.8
Change-Id: Ibdd7e059a2e5e28792ae0fb0a4ca1f133f6e94c2
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>