ForInIteratorObject: also mark target
When dealing with the `for..in` language construct, V4 internally uses a
`ForInIteratorObject` to implement the required iteration.

As part of the initialization of this kind of object, the iterator marks
the elements it iterates upon to avoid them being collect in while the
iteration is still going.

A "proxy" is a kind of object that wraps another object, adding one
layer of indirection.

A "proxy" object can be revocable. Revoking a "proxy" through a revoke
function or the revoke method is generally intended to end its lifetime.

Internally, a `Proxy` object keeps track of its "target" and its related
"handler".

When the `Proxy` object is revoked, its references to those elements are
nulled out and the `Proxy` is unusable.

If the `Proxy` is the only object holding a reference to those elements,
it then means that, after a revoke operation, the elements are eligible
for garbage collection.

It is possible to create such a situation during a `for..in` run over a
revocable `Proxy` object.
If the `Proxy` has the only reference to its "target", and the iterator
doesn't keep the "target" alive itself, then revoking the `Proxy` object
will allow "target" to be collected when it is still in use.

Indeed, while `ForInIteratorObject` has some special handling for
`Proxy` objects, it failed to mark all its dependencies so that the
above situation could be realized.

To fix the issue, ensure that `ForInIteratorObject` marks all relevant
elements when iterating a `Proxy` object.

A test case was added to `tst_qv4mm` that shows the behavior.

Fixes: QTBUG-129310
Pick-to: 6.5
Change-Id: Ic360d77c1495a87153bb57f8ed87d23552edc08b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
(cherry picked from commit 7e150543926c1377475695fdc5a18dbf1f177241)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>