Fix focus out behavior in QQuickComboBox to support popup windows
A QQuickPopup with popupType: Popup.Item will never have active focus
when using a non-custom style. However, a QQuickPopup with popupType:
Popup.Window will always have active focus because it uses a
QQuickWindow.

When the focus property is set to false and a FocusOut event is
triggered in QQuickComboBox, it is first checked that the active focus
of the popup is set to false before hiding the popup. This is so that
custom styles can choose to have active focus and therefore receive key
events. However, this falls apart when using popup windows because they
always have active focus.

This patch adds an addition to the check for active focus in the
FocusOut event, making sure that if popup windows are used, the popup is
still hidden.

Fixes: QTBUG-126987
Pick-to: 6.8
Change-Id: I40b7b4e1e081005545d7c535aba06a490c55fb07
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>