QQuickTableView: ensure VerticalHeaderView stays in sync
Bacuse of a copy-paste error in QQuickTableView, TableView
would calulcate the wrong viewport rect for vertically
synced views if the views were not also synced horizontally.

The result of this mistake was that the rows in the two views
would end up with with different heights (and y pos) while
flicking.

This bug has been there for a very long time, but was hidden
becuse of anther bug that was recently fixed (25348bc0e6).
After that fix, this new bug became evident.

This patch will make sure that the offending if-test checks
the correct sync direction, and thereby also calculate the
correctly combined viewport rect.

Fixes: QTBUG-128508
Pick-to: 6.8 6.7 6.5 6.2
Change-Id: Idf4d93830933c9a28786adaba14e068015564fcc
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>