qmlls documentation hints: fix parsing of html documentation
Also skip the first paragraph if it starts with <p><b>.
The actual documentation usually starts in the second
paragraph if the first paragraph contains the <b>.
This will remove the warnings contained in any accompanying
(\note or \warnings) text box. Also, this post
processing only happens in HtmlExtractor::Simplified mode.

Pick-to: 6.8
Task-number: QTBUG-120016
Change-Id: I367f8aded68972d34f31ed380ed3317865c289f3
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>