QQuickNativeMenuItem: don't check an uncheckable MenuItem
As it stood, some menu items would always end up checked
if triggered. E.g on macOS, when running the textEditor
example, if you open up the context menu and click
"Font...", it would become checked.

This patch will ensure that the MenuItem is infact
checkable before we toggle it.

Task-number: QTBUG-69558
Pick-to: 6.8
Change-Id: Ie71896df4bf706a3e0cba3ef0851e8adae4b15cc
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>