Propagate attached style properties to the Popup with popupType window
The Popup with popupType configured as Popup.Window doesn't propagate
attached style properties. This is because the attached objects (to
handle the attached properties) would normally be created when
specified in QML and any dynamically created object would need to
request QML engine explicitly to make it available.

In the Popup case, the attached object is not readily available as the
QQuickPoupWindow has been created dynamically when popuptype is set as
Window for the Popup (or any item inheriting from it), which causes
the popup window to be stranded with no propagator object to inherit.

This patch sets the popup window's transient parent propagator instance
as its attached parent, if the attached object for the popup window
doesn't already exist.

Fixes: QTBUG-126713
Change-Id: Iaff2a6deef38438823d8d27fddf3036c39df2660
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit d824585075b0abaec12c0cbb6445aa4eb84b2cea)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>