Add a flickDeceleration property to Tumbler
Tumbler is based on PathView and ListView. They both have a
flickDeceleration property. It was not possible to configure this
property when using Tumbler.

flickDeceleration in PathView and ListView does not have the same
default value. Hence, to preserve the original behavior, the property
of Tumbler changes together with the wrap property by default, unless
assigned an explicit value. It is resettable.

The designer file for Tumbler contains the new property as well.

[ChangeLog][Controls][Tumbler] Added flickDeceleration property.

Change-Id: I5f58b8659fea9bbd7bcc7381087351162b8e06c3
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>