QML ComboBox closes on escape/back key release, not press
QQuickComboBox owns a QQuickPopup. When the Popup's close policy is set
to CloseOnEscape, or the default close policy, the Popup closes on key
release, not key press. This contradicts QQuickPopup::ClosePolicy's
documentation that states Popup.CloseOnEscape will close when the escape
key is pressed while the Popup has active focus.

Fixes: QTBUG-128525
Pick-to: 6.8
Change-Id: Ie7a3fbeedba560360f015153ac194a04a038dd0b
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>