QML Previewer Example: Fix bug when saving new file
When saving content as a new file, the application saved the changes but
nothing changed in the quick widget. The problem was that the state was
not changed correctly, and the current state wouldn't transition to the
open state.

To fix this issue, when saving the content, the transition from the new
state to the open state is `changesSaved()`, while `setDirty(false)` was
incorrectly changing the new state back to the init state.

Pick-to: 6.8
Change-Id: I20707046bc17044a18c95ce77e6e56c4a8c21627
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>