Curve renderer: Fix crash when root node is replaced
A curve renderer instance keeps references to the fill and stroke
nodes it has created, in order to facilitate maintaining correct
rendering order when updating. All nodes are children of the
renderer's root node.

It can happen, like when hiding and re-showing a Popup, that the
renderer instance is assigned a new root node. The references to the
child nodes are then invalid, and accessing them will fail.

Fix by clearing all such references when a new root node is set.

Fixes: QTBUG-128561
Pick-to: 6.8 6.7
Change-Id: Ibc4992780867049e9e38a2bf1a163eed79bdc5c0
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>