Make tst_qquickpopup work, independently of the default popupType
Most of the tests in tst_qquickpopup were written before we introduced
the popupType property, and as a result, were making the assumption that
the popupType would always be `Popup.Item`.

Since we plan on changing the default popupType value for desktop OS'es,
we'll need to prepare for that, by making some manual changes to
existing tests, so that existing tests still work as they should.
Since most of them were written with `Popup.Item` in mind, it makes
sense to explicitly force the tests to be ran with that popupType.

Task-number: QTBUG-126553
Change-Id: I2709fd1c416d2ab7877b989c20a70356ce6d5417
Reviewed-by: Jan Arve Sæther <jan-arve.saether@qt.io>
(cherry picked from commit 5097e34f74f46ed144dd80d388efeaa1bbdb747a)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>