Android: Indicate signals with no params with QVariant<void*>
Previously we were sending an invalid QVariant if the emitted
signal had no parameters. Distinction between an invalid
QVariant and a artificial one was not possible in the
Android type converter.

In this change, we create a QVariant from void* that can be
detected as "void" signal during type conversion.

Fixes: QTBUG-128586
Pick-to: 6.8
Change-Id: I6371e25399801c8407e729bfa001ce61dd51dfbb
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Petri Virkkunen <petri.virkkunen@qt.io>