Curve renderer: Avoid losing updates during async processing
If a new sync round, updating path properties, happened while an async
thread was already processing a previous update, it would not trigger
a new round of processing and node update. This would usually not be
noticed, since another sync and update would typically happen shortly
after, say during an animation, and the changes would be picked up
then. However, updates from singular or stopping animations could be
"lost", in the sense that the node (and thus rendering) was not
brought in sync with the updated path properties in a timely fashion.

Fix by kicking off new async processing and update right away when
detecting that path properties have changed while the previous
processing was running.

This also fixes a related minor issue: at node update time, uniforms
where set from the current state, instead of the state as it was when
the path processing started. In the async case, that may be different,
so it was possible for the node properties to be briefly set to an
inconsistent state.

Fixes: QTBUG-128637
Pick-to: 6.8
Change-Id: I44de44c796e8bd7c2f84e80cac0811951b9a2bd4
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>