Fix gc interaction when reparenting items
QQuickItem has sspecial integration code with the garbage collector, to
ensure that an item which is in the (item) parent hierarchy will not get
collected. However, that breaks down with incremental gc in one specific
corner case:
1. We have an item A which already got fully marked (its custom
   markObjects has already run). The phase in which we mark all object
   wrappers has however not ended yet.
2. We have another item B with child item C, and both of their wrappers
   are not marked yet.
3. The itme parent of C changes from B to A.
4. The gc resumes, but won't ever reach C, because it is only reachable
   from A, which already had been visited.

Note that there is an issue in loadFromModule not correctly setting up
the context, which requires a workaround in the test.

Change-Id: Id8d14cd1ea96161d2864f49b43a9369b2f363396
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 75603bfe1af69ca8a9ce45ee37dee722be8b800e)