qmlls: redesign documentation hints
There are two ways to obtain the help through help plugin. The first is
querying the keyword which will probably give a hell of a lot resulting
results to be processed. Better way is to search the keyword by id which
will give the precise documentation for the given name. id's are kept in
the form of DocumentName::keyword form in the relavant .qhp file.

So, to get the correct documentation for an identifier, one should find
the defining scope for that identifier and try to get the relavant link
by searching DefiningScopeName::identifierName.

Documentation for most of the qml properties are found in the Qml named
namespace in the .qhp file. For example, x,y,width,height of an Item are
defined in QQuickItem scope but documentation doesn't exist for
QQuickItem::x, but for Item::x. To get the proper mapping, utilise
typeResolver->importedNames() method and introduce the m_cppTypesToQmlTypes
hash map.

Task-number: QTBUG-127051
Task-number: QTBUG-120016
Change-Id: I27d4c60b11cd2aff4b4cb8a19ab9817519799e06
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
(cherry picked from commit ffd39a4f59e5cd57a731549395218d9c96e2bea2)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>