QmlCompiler: Fix enum scoping rules
You can always access a scoped enum in a scoped way in QML, no matter if
the type it belongs to enforces scoped enum access. Read the respective
flag separately and don't mix it in with the scopedness of the actual
enums.

Amends commit 7da544e862a92cc0a9d97cbed68a35c532699116

Pick-to: 6.8 6.5 6.2
Fixes: QTBUG-127308
Change-Id: I167602d44c8cea874dc39614fc6726a72e66c007
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>