Care about signal ordering in popupWindowDestructedBeforeQQuickPopup()
tst_qquickpopup::popupWindowDestructedBeforeQQuickPopup() verifies that
the popup window is deleted upon the destruction of the popup object.

Since both objects are destructed in the same frame, and almost
simultaneously, the QTRY_COMPAREs weren't sufficient in making sure that
the objects are deleted in the correct order.

An existing comment suggest that the popup window should be destructed
after the popup object, but I believe that's false, since the popup
window is simply destructed in the popup's destructor.
With this in mind, it makes sense to update the comment to instead
mention that we expect the popup window to be destructed first.

We can verify that the ordering is correct, by connecting a lambda to
the popup windows destroyed signal, and check the signalspy for the
popup object.

Pick-to: 6.8 6.8.0
Change-Id: I3a40926795978b5fbda3478a6bedc691dc8ae162
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Ed Cooke <ed.cooke@qt.io>