Doc: Clarify the use of qsTranslate()
There are two QT QML types that can be used for translations, namely,
qsTr() and qsTranslate(). This change adds a description to the docs to
help the user determine when to use qsTranslate() over qsTr().

qsTranslate() takes an explicit context as its first argument, whereas
qsTr() takes the text to be translated as its first argument. If the
user has a translation context that differs from the file context for
some strings, then it would be better to use qsTranslate() for those
strings.

Fixes: QTBUG-127796
Pick-to: 6.7 6.8
Change-Id: Ic7d44228f293666c112ecf47cf989ce2617f71e0
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>