MenuBar: don't reopen a menu if we intend to close all popups
If you click outside a menu bar menu, all menus will be
closed by QQuickPopupWindow. But such clicks will also be
forwarded to the window under the mouse (by QGuiApplication).
The result is that if you click on top of the menu bar to
close the current menu, you will also end up clicking on a
MenuBarItem, which therefore reopens the menu again.

To solve this bug, we let forwardEventToParentMenuOrMenuBar()
return a bool that tells if the event should be blocked from
further propagation or not. That way we can return true if we
detect that the click is on top the menu bar, and should not
be forwarded down to it if we also use it to close all menus.

Pick-to: 6.8.0
Change-Id: I9ba638801218393528536c49a186ae5b2c190a5a
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 129a5d606ad191bd7bb312bad421ff5f979cc14b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>