QQuickMenu: always use native menu if MenuBar is native
If the MenuBar is native, all menus inside it needs to be
native as well, even if AA_DontUseNativeMenuWindows is set.
If the application don't want a native menu bar, it needs
to additionally set AA_DontUseNativeMenuBar.

As it stood, we QQuickMenu would refuse to create a native
menu for a native MenuBar if AA_DontUseNativeMenuWindows
was set. And this would lead to an assert in
QQuickMenuPrivate::nativeHandle().

This patch will therefore change this, so that we respect
AA_DontUseNativeMenuWindows only if the MenuBar is not
already native.

Pick-to: 6.8 6.8.0
Change-Id: I4b8bc0ab8bf409f3a6c755e70223df3e042f8220
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>