qmlformat: refactor qmlformat.ini reading
Move the Options to QQmlFormatOptions and make it reuse the fields
from LineWriteOptions via setters and getters instead of duplicating them.

Create a QQmlFormatSettings class that is a QQmlToolingSettings with
qmlformat fields, and that contains the static keys of the settings it
contains.

Those classes will be moved to a separate library and reused by qmlls
in a later commit.

Task-number: QTBUG-128866
Change-Id: I90057fedbae57e487e64fb98181ae26730008f67
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>