QQuickFocusFrame: Fix showing focus frame inside embedded qquickwidget
In QQuickFocusFrame, we connect to QGuiApplication::focusObjectChanged
signal and draw the focus frame around focusObject() of qGuiApp, which
as per its implementation, is the focusObject() of the app's
focusWindow(). However, in the case of QQuickWidget, its
offscreenWindow will never become the focusWindow of the application
because it doesn't have a native handle. Also, since
0dbc575c1a8359534761167a5f5f1e29abedd51d QQuickWidget intentionally
doesn't override the focusObject method to return the focusObject of
its offscreen window because that would bypass needed event handling
logic in qquickwidget.

That is why if a qquickwidget has active focus, qGuiApp will report
the qquickwidget itself as its focusObject, rather than the focusObject
inside of the widget. However, we still do propagate focus object
changes inside the qquickwdiget's offscreen window up to the top
level widget and the application, so that when the offscreen's window
QQuickWindow::focusObjectChanged is emitted, the equivalent signal
in QGuiApplication will also be emitted.
That is why in qquickfocusframe we should rely on the signal parameter
instead, which represents the right focus object, rather than query
qGuiApp::focusObject().
While we're at it, reset the focus frame if no item inside the quick
window currently has active focus, otherwise we would still show
a focus frame even if the quick window lost focus.

Note: Ideally qGuiApp::focusObject() should give us the same object as
the QGuiApplication::focusObjectChanged signal parameter, but that
can require more work as the focusObject() of the application will
then receive key events and qquickwidget relies on handling those key
(and other) events itself.

Fixes: QTBUG-128931
Pick-to: 6.8.0
Change-Id: Id86a4e9dfd6b2bda5cc9ea633305e71d86fcdb5f
Reviewed-by: MohammadHossein Qanbari <mohammad.qanbari@qt.io>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 9a83ce4ebaab1cbaa6915ebecef43147d4ec55e7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>