Use QPointer for QQuickPixmapData::specialDevice; abort if deleted
QQuickPdfPageImage::load() calls carrierFile() to get a QPdfFile
instance, a subclass of QIODevice, and calls
QQuickPixmap::loadImageFromDevice() which saves the device to
QQuickPixmapData::specialDevice, then calls
QQuickPixmapReader::startJob() which sends a ProcessJobs event.
The worker thread handles it in QQuickPixmapReader::processJob()
and calls readImage(), passing a simple QIODevice *dev. readImage()
relies on that pointer being valid until the work is done. However,
QQuickPdfDocument::setSource() may have called deleteLater().
We can detect it, and in that case there is no point in continuing this
render job. Also, change QPdfFile's thread affinity to the renderer
thread, so that deletion cannot happen until after readImage() is done
and the renderer thread has returned to its event loop.

Fixes: QTBUG-128875
Pick-to: 6.8 6.7 6.5
Change-Id: I1e8bf526c5596ce561d7bf4f7dea8d50f3c0eb18
Reviewed-by: Axel Spoerl <axel.spoerl@qt.io>