Try verify scheduled polish in the table view for resize test case
In the table view auto test, the section resize was performed through
mouse events (press, move and release operation). The scheduled polish
is validated immediately after those events before checking for the
change in section sizes. This can cause the resize test cases to be
flaky, as there is a chance that the scheduled polish would have
updated the view even before we assumed that the polish had been
scheduled.

This patch changes the condition by try verifying the scheduled polish.

Fixes: QTBUG-128934
Pick-to: 6.8
Change-Id: I6520fbbca4e34b4f47035b36f8c92f77431d82a6
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>