QtQml: Decouple JavaScript library CUs from engines
In order to re-use the compilation units for JavaScript libraries, we
need to eliminate the "m_value" member they are carrying, indirectly.
The values are tied to specific engines and invalid in others.

Luckily, we already have two suitable places to store such values:
1. In case of a "native" module without a compilation unit we have the
   nativeModules hash in ExecutionEngine.
2. In case of a module or library backed by a CU we have the "module"
   member of ExecutableCompilationUnit. This can currently only hold
   modules but there is no reason why it wouldn't hold JavaScript
   libraries equally well.

By using the "empty" V4 value we can also get rid of the m_loaded bool.
As a drive by, correct the QQmlScriptBlob::isNative() misnomer. We don't
want to know whether the script is native (for any value of that), but
rather whether it has a value.

Pick-to: 6.8
Fixes: QTBUG-129052
Change-Id: I284823f4aa26e46dec8328f88f65877f59e40f79
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>