Re-enable Fbo mode in QQuickPaintedItem with OpenGL only
For the time being this is feasible since the Qt Quick library links to
(Qt)OpenGL still due to classes such as QQuickFbo or QQuickOpenGLUtils.
In principle, having OpenGL-specific code returning to the common code
base is far from ideal, but since Qt Quick is not entirely free from
this anyway, we might just live with it for now.

Import the Qt 5 code in the painter node, all under ifdef
QT_CONFIG(opengl) and a runtime check for the QRhi backend.

[ChangeLog][QtQuick][QQuickPaintedItem] QQuickPaintedItem is made
to behave identically to Qt 5 when the application is rendering with
OpenGL and the renderTarget property is set to FramebufferObject or
InvertedYFramebufferObject.

Change-Id: If5297368836ca8373aae0365c5ae00a7b1ad50d9
Reviewed-by: Andy Nichols <andy.nichols@qt.io>