tests: Explicitly check for pathconf(_PC_CASE_SENSITIVE) == 1
We use pathconf in some of our tests to determine if the file system
on Darwin is case sensitive. But pathconf returns -1 if the case
sensitivity can't be determined, with errno set to ENOTSUP.

The convention in this case is to treat the file system as not being
case sensitive (as reported by NSURLVolumeSupportsCaseSensitiveNamesKey
and VOL_CAP_FMT_CASE_SENSITIVE in equivalent APIs), so we need to check
explicitly for a return value of 1.

Change-Id: I692e396bbfddfd0754a89dff2207588a10627579
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Liang Qi <liang.qi@qt.io>
(cherry picked from commit 77ddf924b16af476d1b8590306836d8ccc9c87f9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>