Fix flaky test case in the quick table view
The test case added as part of the sha
2cfffb708808ba49148ba4f30c7c8841e5dfb2d9 can be flaky. The validation
for the scheduled polish seems to be causing this issue. Its because the
view is polished even before we verify for scheduled polish. Remove that
existing validation and make sure the table view loads the required rows
with the timeout.

Amends patch 2cfffb708808ba49148ba4f30c7c8841e5dfb2d9.

Pick-to: 6.8 6.7 6.5
Change-Id: I9a8c5874bf982bc6245cb1ca531fe4399d38b13a
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>