QML Previewer Example: Fix saving the default file's changes
When changes to the default loaded file were saved, the address bar
displayed an incorrect path and the view failed to update.

The issue stemmed from an improper transition from the Dirty state to
the Open state. The correct transition is changesSaved(). The
setDirty(false) method is reserved for when changes are discarded (e.g.,
after an undo operation).

Pick-to: 6.8 6.8.0
Change-Id: Iafee4bf7c0eeae6f2fc7fdb0d32a555a99824249
Reviewed-by: Doris Verria <doris.verria@qt.io>