QtQml: Allow creating attachments from unregistered types
We can resolve the attached properties function without querying the
type registry. The attached properties function can be resolved with
pure C++, given it's surrounding type, name, and signature. We already
have functionality to do so in qqmlprivate.h

This is not only faster, but also avoids the problem of having to
register dependent types before being able to use their attachments.

Pick-to: 6.8 6.5
Fixes: QTBUG-128895
Change-Id: I5c1df81a43195eb4dd97a4abfaa79f3cfe39355e
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>