macOS: Don't trust canonical file path on non-case-preserving FS
APFS and HFS+ are both case preserving, in which case we can trust
that the canonical file path reported above is correct.

But if any of these file systems are mounted via SMB or Virtiofs
(via Virtualization.framework) macOS will report that the canonical
file name of "Foo" is "Foo", even if the underlying file on the host
file system has a canonical name of "foo".

This caused test failures when running tests on a VM with the
build and source directories mounted from the host via Virtiofs
or SMB, as we would pick up date.qml as if it was Date.qml, which
ended up shadowing the built-in JS Date type.

As we can't trust the canonical name in this case, we go though
QDirIterator instead, as the directory listing for the mounted
filesystem _does_ report the correct canonical file name.

As a drive-by optimization we also treat QRC paths as being the
correct file name, just like we do on Windows, as the QRC file
system is case sensitive.

Change-Id: Ic16c1685268afb893887719a0f012d01757e1052
Reviewed-by: Anton Kudryavtsev <antkudr@mail.ru>
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>