Revert "QmlCompiler: Use lookupResultMetaType() also for calls"
This reverts commit 1d59171d19453fb67df2b2bb70b847b6d3300574.

There are a number of places where we pass invalid metatypes to the
various lookup methods in place of metatypes for QObject-derived
pointers. Miraculously this actually works because we can determine the
type of QObject pointers via run time inspection.

We cannot replace the "not initialized" type with an explicit one
before all of those miracles are de-mystified.

Fixes: QTBUG-129202
Change-Id: I84e4d769dc00ef6ebdfcdc9717e30c17e30e811e
Reviewed-by: Jarkko Koivikko <jarkko.koivikko@code-q.fi>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>