QQuickMenu: refactor all 'use native' logic into a single function
As it stood, the logic that resolved if a menu can, and
should be, native was duplicated into two functions;
useNativeMenu() and resolvedPopupType(). And the two was
somewhat inconsistent.

This patch will therefore refactor all the logic into
resolvedPopupType(), so that the popup type is resolved
in a single place. Existing auto tests
(such as tst_QQuickMenu::subMenuPopupType()) should verify
that this refactoring is safe.

Change-Id: Ifc73d20c2d2e06646dc2849c8398660a35b17cb2
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
(cherry picked from commit 6623e7b322bf427c27c0a52723376d0e7e174653)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>